+++ Makefile.toml
@@ -125,7 +125,7 @@ args = [
 
 [tasks.check]
 description = "Runs rustfmt and clippy checks without applying any fix"
-dependencies = ["check-rustfmt", "check-clippy"]
+dependencies = ["check-clippy", "check-rustfmt"]
 
 [tasks.check-rustfmt]
 description = "Runs rustfmt checks without applying any fix"
@@ -197,7 +197,7 @@ args = [
 ]
 
 [tasks.worker-executor-tests-group1]
-description = "Runs worker executor tests only (group 1/6)"
+description = "Runs worker executor tests only (group 1/8)"
 env = { "RUST_BACKTRACE" = "1", "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
 script = '''
 cargo test --package golem-worker-executor-base --test integration api -- --nocapture
@@ -206,14 +206,14 @@ cargo test --package golem-worker-executor-base --test integration keyvalue -- -
 '''
 
 [tasks.worker-executor-tests-group2]
-description = "Runs worker executor tests only (group 2/6)"
+description = "Runs worker executor tests only (group 2/8)"
 env = { "RUST_BACKTRACE" = "1", "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
 script = '''
 cargo test --package golem-worker-executor-base --test integration guest_languages -- --nocapture
 '''
 
 [tasks.worker-executor-tests-group3]
-description = "Runs worker executor tests only (group 3/6)"
+description = "Runs worker executor tests only (group 3/8)"
 env = { "RUST_BACKTRACE" = "1", "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
 script = '''
 cargo test --package golem-worker-executor-base --test integration transactions -- --nocapture
@@ -221,7 +221,7 @@ cargo test --package golem-worker-executor-base --test integration wasi -- --noc
 '''
 
 [tasks.worker-executor-tests-group4]
-description = "Runs worker executor tests only (group 4/6)"
+description = "Runs worker executor tests only (group 4/8)"
 env = { "RUST_BACKTRACE" = "1", "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
 script = '''
 cargo test --package golem-worker-executor-base --test integration scalability -- --nocapture
@@ -231,7 +231,7 @@ cargo test --package golem-worker-executor-base --test integration rust_rpc -- -
 '''
 
 [tasks.worker-executor-tests-group5]
-description = "Runs worker executor tests only (group 5/6)"
+description = "Runs worker executor tests only (group 5/8)"
 env = { "RUST_BACKTRACE" = "1", "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
 script = '''
 cargo test --package golem-worker-executor-base --test integration guest_languages2 -- --nocapture
@@ -239,10 +239,24 @@ cargo test --package golem-worker-executor-base --test integration guest_languag
 
 
 [tasks.worker-executor-tests-group6]
-description = "Runs worker executor tests only (group 6/6)"
+description = "Runs worker executor tests only (group 6/8)"
 env = { "RUST_BACKTRACE" = "1", "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
 script = '''
 cargo test --package golem-worker-executor-base --test integration ts_rpc -- --nocapture
+'''
+
+[tasks.worker-executor-tests-group7]
+description = "Runs worker executor tests only (group 7/8)"
+env = { "RUST_BACKTRACE" = "1", "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
+script = '''
+cargo test --package golem-worker-executor-base --test integration guest_languages3 -- --nocapture
+'''
+
+[tasks.worker-executor-tests-group8]
+description = "Runs worker executor tests only (group 8/8)"
+env = { "RUST_BACKTRACE" = "1", "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
+script = '''
+cargo test --package golem-worker-executor-base --test integration ts_rpc2 -- --nocapture
 '''
 
 [tasks.integration-tests]
