+++ golem-worker-executor-base/tests/ts_rpc.rs
@@ -86,151 +86,3 @@ async fn counter_resource_test_1_with_restart() {
     check!(result1 == Ok(vec![Value::U64(1)]));
     check!(result2 == Ok(vec![Value::U64(2)]));
 }
-
-#[tokio::test]
-#[tracing::instrument]
-async fn counter_resource_test_2() {
-    let context = common::TestContext::new();
-    let executor = common::start(&context).await.unwrap();
-
-    let counters_component_id = executor.store_component(COUNTER_COMPONENT_NAME).await;
-    let caller_component_id = executor.store_component(CALLER_COMPONENT_NAME).await;
-
-    let mut env = HashMap::new();
-    env.insert(
-        "COUNTERS_COMPONENT_ID".to_string(),
-        counters_component_id.to_string(),
-    );
-    let caller_worker_id = executor
-        .start_worker_with(&caller_component_id, "rpc-counters-2", vec![], env)
-        .await;
-
-    let result1 = executor
-        .invoke_and_await(&caller_worker_id, "test2", vec![])
-        .await;
-    let result2 = executor
-        .invoke_and_await(&caller_worker_id, "test2", vec![])
-        .await;
-
-    drop(executor);
-
-    check!(result1 == Ok(vec![Value::U64(1)]));
-    check!(result2 == Ok(vec![Value::U64(2)]));
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn counter_resource_test_2_with_restart() {
-    let context = common::TestContext::new();
-    let executor = common::start(&context).await.unwrap();
-
-    let counters_component_id = executor.store_component(COUNTER_COMPONENT_NAME).await;
-    let caller_component_id = executor.store_component(CALLER_COMPONENT_NAME).await;
-
-    let mut env = HashMap::new();
-    env.insert(
-        "COUNTERS_COMPONENT_ID".to_string(),
-        counters_component_id.to_string(),
-    );
-    let caller_worker_id = executor
-        .start_worker_with(&caller_component_id, "rpc-counters-2r", vec![], env)
-        .await;
-
-    let result1 = executor
-        .invoke_and_await(&caller_worker_id, "test2", vec![])
-        .await;
-
-    drop(executor);
-    let executor = common::start(&context).await.unwrap();
-
-    let result2 = executor
-        .invoke_and_await(&caller_worker_id, "test2", vec![])
-        .await;
-
-    drop(executor);
-
-    check!(result1 == Ok(vec![Value::U64(1)]));
-    check!(result2 == Ok(vec![Value::U64(2)]));
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn context_inheritance() {
-    let context = common::TestContext::new();
-    let executor = common::start(&context).await.unwrap();
-
-    let counters_component_id = executor.store_component(COUNTER_COMPONENT_NAME).await;
-    let caller_component_id = executor.store_component(CALLER_COMPONENT_NAME).await;
-
-    let mut env = HashMap::new();
-    env.insert(
-        "COUNTERS_COMPONENT_ID".to_string(),
-        counters_component_id.to_string(),
-    );
-    env.insert("TEST_CONFIG".to_string(), "123".to_string());
-    let caller_worker_id = executor
-        .start_worker_with(
-            &caller_component_id,
-            "rpc-counters-4",
-            vec!["a".to_string(), "b".to_string(), "c".to_string()],
-            env,
-        )
-        .await;
-
-    let result = executor
-        .invoke_and_await(&caller_worker_id, "test3", vec![])
-        .await;
-
-    drop(executor);
-
-    let result = result.unwrap();
-    let result_tuple = match &result[0] {
-        Value::Tuple(result) => result,
-        _ => panic!("Unexpected result: {:?}", result),
-    };
-    let args = match &result_tuple[0] {
-        Value::List(args) => args.clone(),
-        _ => panic!("Unexpected result: {:?}", result),
-    };
-    let mut env = match &result_tuple[1] {
-        Value::List(env) => env
-            .clone()
-            .into_iter()
-            .map(|value| match value {
-                Value::Tuple(tuple) => match (&tuple[0], &tuple[1]) {
-                    (Value::String(key), Value::String(value)) => (key.clone(), value.clone()),
-                    _ => panic!("Unexpected result: {:?}", result),
-                },
-                _ => panic!("Unexpected result: {:?}", result),
-            })
-            .collect::<Vec<_>>(),
-        _ => panic!("Unexpected result: {:?}", result),
-    };
-    env.sort_by_key(|(k, _v)| k.clone());
-
-    check!(
-        args == vec![
-            Value::String("a".to_string()),
-            Value::String("b".to_string()),
-            Value::String("c".to_string())
-        ]
-    );
-    check!(
-        env == vec![
-            (
-                "COUNTERS_COMPONENT_ID".to_string(),
-                counters_component_id.to_string()
-            ),
-            (
-                "GOLEM_COMPONENT_ID".to_string(),
-                counters_component_id.to_string()
-            ),
-            ("GOLEM_COMPONENT_VERSION".to_string(), "0".to_string()),
-            (
-                "GOLEM_WORKER_NAME".to_string(),
-                "counters_test4".to_string()
-            ),
-            ("TEST_CONFIG".to_string(), "123".to_string())
-        ]
-    );
-}
